refactor and remove unused code

+++ src/main/kotlin/acoflowshop/CsvLogging.kt
@@ -1,10 +1,15 @@
 package acoflowshop
 
+import com.fasterxml.jackson.databind.ObjectMapper
+import com.fasterxml.jackson.module.kotlin.KotlinModule
+import global.Config
 import mu.KotlinLogging
 import java.io.File
 
-private val FILE_NAME = "current"
+private const val FILE_NAME = "current"
 private val logger = KotlinLogging.logger {}
+private val mapper = ObjectMapper().registerModule(KotlinModule())
+private val config = mapper.readValue(File("src/main/resources/Config.json"), Config::class.java)
 
 class CsvLogging {
 
@@ -13,8 +18,10 @@ class CsvLogging {
          * löschen des Logs der letzten iteration und erstellen eines neuen Files
          */
         fun createLoggingFile() {
-            File("${FILE_NAME}.csv").delete()
-            File("${FILE_NAME}.csv").createNewFile()
+            if(config !== null && config.fileLogging) {
+                File("${FILE_NAME}.csv").delete()
+                File("${FILE_NAME}.csv").createNewFile()
+            }
         }
 
         /**
@@ -22,12 +29,16 @@ class CsvLogging {
          */
         fun appendCSVEntry(iteration: Int, currentLength: Int, durationInMs: Long) {
             logger.info { "${iteration} - ${currentLength} - ${durationInMs}" }
-            File("${FILE_NAME}.csv").appendText("${iteration},${currentLength},${durationInMs}\n")
+            if(config !== null && config.fileLogging) {
+                File("${FILE_NAME}.csv").appendText("${iteration},${currentLength},${durationInMs}\n")
+            }
         }
 
         fun appendCSVEntry(iteration: Int, currentLength: Double, durationInMs: Long) {
             logger.info { "${iteration} - ${currentLength} - ${durationInMs}" }
-            File("${FILE_NAME}.csv").appendText("${iteration},${currentLength},${durationInMs}\n")
+            if(config !== null && config.fileLogging) {
+                File("${FILE_NAME}.csv").appendText("${iteration},${currentLength},${durationInMs}\n")
+            }
         }
     }
 }
